requirements.txt
cuid2==2.0.1
deepgram-sdk==3.4.0
ibmcloudant==0.9.1
mutagen==1.47.0
boto3==1.34.149
openai==1.37.1